refactor server buildServerTLSConfig function into tls package
