+++ rest/server.go
@@ -2,7 +2,6 @@ package rest
 
 import (
 	"context"
-	"crypto/tls"
 	"fmt"
 	"net"
 	"net/http"
@@ -12,7 +11,7 @@ import (
 	"sync"
 	"syscall"
 
-	"github.com/fsnotify/fsnotify"
+	"github.com/cloudhut/common/tls"
 	"github.com/go-chi/chi/v5"
 	"go.uber.org/zap"
 )
@@ -42,7 +41,7 @@ func NewServer(cfg *Config, logger *zap.Logger, router *chi.Mux) (*Server, error
 	}
 
 	if cfg.TLS.Enabled {
-		tlsCfg, err := buildServerTLSConfig(logger, cfg.TLS)
+		tlsCfg, err := tls.BuildWatchedTLSConfig(logger, cfg.TLS.CertFilepath, cfg.TLS.KeyFilepath)
 		if err != nil {
 			return nil, fmt.Errorf("failed to create TLS config: %w", err)
 		}
@@ -108,94 +107,4 @@ func (s *Server) Start() error {
 	s.Logger.Info("Stopped HTTP server", zap.String("address", listener.Addr().String()), zap.Int("port", listenerPort))
 
 	return nil
-}
-
-func buildServerTLSConfig(logger *zap.Logger, cfg TLSConfig) (*tls.Config, error) {
-	cert, err := tls.LoadX509KeyPair(cfg.CertFilepath, cfg.KeyFilepath)
-	if err != nil {
-		return nil, fmt.Errorf("failed loading TLS cert: %w", err)
-	}
-
-	signalCh := make(chan os.Signal, 1)
-	signal.Notify(signalCh, syscall.SIGINT, syscall.SIGTERM)
-
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		return nil, fmt.Errorf("failed to setup file watcher for hot reloading tls certificates: %w", err)
-	}
-
-	var lock sync.RWMutex
-	go func() {
-		defer watcher.Close()
-		for {
-			select {
-			case event, ok := <-watcher.Events:
-				if !ok {
-					return
-				}
-
-				// Ignore all events that are neither remove nor write
-				isRelevantEvent := event.Has(fsnotify.Remove) || event.Has(fsnotify.Write)
-				if !isRelevantEvent {
-					continue
-				}
-
-				if event.Has(fsnotify.Remove) {
-					// Kubernetes uses symbolic links to create the illusion of atomic writes.
-					// Thus, we have to watch for the remove event and reconfigure our watcher.
-					// See: https://ahmet.im/blog/kubernetes-inotify/
-					_ = watcher.Remove(event.Name)
-					err = watcher.Add(cfg.CertFilepath)
-					if err != nil {
-						logger.Error("failed to re-add file watcher",
-							zap.String("file_path", cfg.CertFilepath),
-							zap.Error(err))
-					}
-
-					err = watcher.Add(cfg.KeyFilepath)
-					if err != nil {
-						logger.Warn("failed to re-add file watcher",
-							zap.String("file_path", cfg.KeyFilepath),
-							zap.Error(err))
-					}
-				}
-
-				logger.Info("hot reloading the TLS certificate")
-
-				newCert, err := tls.LoadX509KeyPair(cfg.CertFilepath, cfg.KeyFilepath)
-				if err != nil {
-					logger.Error("failed to load certificates", zap.Error(err))
-					continue
-				}
-				lock.Lock()
-				cert = newCert
-				lock.Unlock()
-
-				logger.Info("successfully hot reloaded the TLS certificate")
-			case err, ok := <-watcher.Errors:
-				if !ok {
-					return
-				}
-				logger.Error("tls certificate watcher error", zap.Error(err))
-			case <-signalCh:
-				return
-			}
-		}
-	}()
-	if err := watcher.Add(cfg.CertFilepath); err != nil {
-		return nil, fmt.Errorf("failed to setup watcher for cert file: %w", err)
-	}
-	if err = watcher.Add(cfg.KeyFilepath); err != nil {
-		return nil, fmt.Errorf("failed to setup watcher for key file: %w", err)
-	}
-
-	return &tls.Config{
-		MinVersion: tls.VersionTLS12,
-		GetCertificate: func(info *tls.ClientHelloInfo) (*tls.Certificate, error) {
-			lock.RLock()
-			defer lock.RUnlock()
-
-			return &cert, nil
-		},
-	}, nil
 }
